feat: 좋아요 기능 추가, oauth이용한 로그인 세션 작업 완료 등

+++ pages/api/post/delete.js
@@ -1,13 +1,22 @@
 import { connectDB } from "@/util/database";
 import { ObjectId } from "mongodb";
+import { getServerSession } from "next-auth";
+import { authOptions } from "../auth/[...nextauth]";
 
 export default async function handler(req, res) {
   if (req.method == "DELETE") {
+    let session = await getServerSession(req, res, authOptions);
     let db = (await connectDB).db("yjproject");
     let result = await db
       .collection("post")
-      .deleteOne({ _id: new ObjectId(req.body) });
-    // console.log(result);
-    res.status(200).json("삭제완료");
+      .findOne({ _id: new ObjectId(req.body) });
+    if (result.author == session.user.email) {
+      let result = await db
+        .collection("post")
+        .deleteOne({ _id: new ObjectId(req.body) });
+      return res.status(200).json("삭제완료");
+    } else {
+      return res.status(500).json("현재유저와 작성자 불일치");
+    }
   }
 }
